+++ apollo-openapi/src/main/java/com/ctrip/framework/apollo/openapi/client/ApolloOpenApiClient.java
@@ -22,6 +22,7 @@ import com.ctrip.framework.apollo.openapi.client.service.ClusterOpenApiService;
 import com.ctrip.framework.apollo.openapi.client.service.ItemOpenApiService;
 import com.ctrip.framework.apollo.openapi.client.service.NamespaceOpenApiService;
 import com.ctrip.framework.apollo.openapi.client.service.ReleaseOpenApiService;
+import com.ctrip.framework.apollo.openapi.client.service.InstanceOpenApiService;
 import com.ctrip.framework.apollo.openapi.client.service.OrganizationOpenApiService;
 import com.ctrip.framework.apollo.openapi.dto.*;
 import com.google.common.base.Preconditions;
@@ -51,6 +52,7 @@ public class ApolloOpenApiClient {
   private final ReleaseOpenApiService releaseService;
   private final NamespaceOpenApiService namespaceService;
   private final ClusterOpenApiService clusterService;
+  private final InstanceOpenApiService instanceService;
   private static final Gson GSON = new GsonBuilder().setDateFormat(ApolloOpenApiConstants.JSON_DATE_FORMAT).create();
 
   private ApolloOpenApiClient(String portalUrl, String token, RequestConfig requestConfig) {
@@ -66,6 +68,7 @@ public class ApolloOpenApiClient {
     itemService = new ItemOpenApiService(client, baseUrl, GSON);
     releaseService = new ReleaseOpenApiService(client, baseUrl, GSON);
     organizationOpenService = new OrganizationOpenApiService(client, baseUrl, GSON);
+    instanceService = new InstanceOpenApiService(client, baseUrl, GSON);
   }
 
   public void createApp(OpenCreateAppDTO req) {
@@ -241,6 +244,14 @@ public class ApolloOpenApiClient {
    */
   public void rollbackRelease(String env, long releaseId, String operator) {
     releaseService.rollbackRelease(env, releaseId, operator);
+  }
+
+  /**
+   * Get instance count by namespace
+   * @since 2.5.0
+   */
+  public int getInstanceCountByNamespace(String appId, String env, String clusterName, String namespaceName) {
+    return instanceService.getInstanceCountByNamespace(appId, env, clusterName, namespaceName);
   }