+++ CHANGES.md
@@ -7,6 +7,7 @@ Apollo Java 2.5.0
 ------------------
 
 * [Feature Provide a new open APl to return the organization list](https://github.com/apolloconfig/apollo-java/pull/102)
+* [Feature Added a new feature to get instance count by namespace.](https://github.com/apolloconfig/apollo-java/pull/103)
 
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo-java/milestone/5?closed=1)